pyproject.toml
[tool.poetry]
name = "note-transcription"
version = "0.1.0"
description = ""
authors = ["Sundance"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-genai = "^1.2.0"
ipykernel = "^6.29.5"
tqdm = "^4.67.1"
pillow = "^11.1.0"
pydantic = "^2.10.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
